move enableWhen to seperate file
